Change to xspress3 with extra channels for MCA plus PR feedback
