+++ src/dodal/devices/xspress3/xspress3_channel.py
@@ -10,13 +10,17 @@ class TimeSeriesValues(str, Enum):
     UPDATE_VALUE = ""
 
 
-class Xspress3MiniChannel(Device):
+class Xspress3Channel(Device):
+    """
+    Xspress3 Channel contains the truncated detector data and its collection condition
+     including the definition of ROI.
+    """
+
     def __init__(self, prefix: str, name: str = "") -> None:
-        # Define some signals
         self.update_arrays = epics_signal_rw(float, prefix + "SCAS:TS:TSAcquire")
 
         self.roi_high_limit = epics_signal_rw(float, prefix + "SCA5_HLM")
-        self.roi_llm = epics_signal_rw(float, prefix + "SCA5_LLM")
+        self.roi_low_limit = epics_signal_rw(float, prefix + "SCA5_LLM")
 
         self.time = epics_signal_r(float, prefix + "SCA0:Value_RBV")
         self.reset_ticks = epics_signal_r(float, prefix + "SCA1:Value_RBV")
@@ -25,4 +29,11 @@ class Xspress3MiniChannel(Device):
         self.all_good = epics_signal_r(float, prefix + "SCA4:Value_RBV")
         self.pileup = epics_signal_r(float, prefix + "SCA7:Value_RBV")
         self.total_time = epics_signal_r(float, prefix + "SCA8:Value_RBV")
+        super().__init__(name=name)
+
+
+class Xspress3ROIChannel(Device):
+    def __init__(self, prefix: str, name: str = "") -> None:
+        self.roi_start_x = epics_signal_rw(float, prefix + "MinX")
+        self.roi_size_x = epics_signal_rw(float, prefix + "SizeX")
         super().__init__(name=name)
